do not clearTempColors in fillTempColorInClosedArea

+++ fdkgametest/theta/Types.h
@@ -111,9 +111,9 @@ inline int getPlotActorRadiusInVertex()
 	return 30*MAX_UNIT_SIZE*2;
 }
 
-inline int getPlotActorRadius()
+inline float getPlotActorRadius()
 {
-	return getPlotActorRadiusInVertex()*HALF_CELL_SIZE_X;
+	return (float)getPlotActorRadiusInVertex()*HALF_CELL_SIZE_X;
 }
